Move `seed_trait_files` to `db_operations` module
I wasn't originally sure how to do the local import, but it turns out it's very easy :)
